SDB/src/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
numpy
jpholiday